requirements.txt
streamlit
langchain
langchain-google-genai
chromadb
cohere
python-dotenv